config: split source model
