+++ crates/hooks/src/source_switch.rs
@@ -59,10 +59,7 @@ pub async fn apply_source_switch(
 ) -> bool {
     match source {
         Source::Local => {
-            let mut cfg = config.write();
-            cfg.active_source = Source::Local;
-            cfg.server = None;
-            cfg.source_explicitly_set = true;
+            config.write().clear_active_server();
             tracing::info!(target: "kopuz::source", source = "local", "source switched");
             true
         }
@@ -95,9 +92,7 @@ pub async fn apply_source_switch(
             };
             {
                 let mut cfg = config.write();
-                cfg.active_source = Source::Server(saved.id);
-                cfg.server = Some(active);
-                cfg.source_explicitly_set = true;
+                cfg.set_active_server_snapshot(active);
             }
             tracing::info!(target: "kopuz::source", server = %id, "source switched");
             has_creds || is_anon
